Display Item Name and Price when found.
This felt like WAY too big of a step. This took a long time in the red and I had multiple failing tests as I began to get the new test passing.

This happens to me a lot when I try TDD.

I don't like overriding ToString. I'd want a PresentableItem that can handle this instead of doing it in the business entity.